% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_curve.R
\name{discrete_curve}
\alias{discrete_curve}
\title{Iterate through result folders to obtain a dataframe}
\usage{
discrete_curve(gd, resolution = 100)
}
\arguments{
\item{gd}{The output of a \code{plot_geodist()} call}

\item{resolution}{numerical How many bins should the result have?}
}
\value{
A dataframe
}
\description{
Discretizes the dataframe output by \code{plot_geodist()}.
}
