% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMD.R
\name{EMD}
\alias{EMD}
\title{EMD between sample-to-prediction and CV-distances}
\usage{
EMD(df, dist1, dist2)
}
\arguments{
\item{df}{A dataframe, output of \code{emodis::discrete_curve()}}

\item{dist1}{The name of the first distribution}

\item{dist2}{The name of the second distributions}
}
\value{
The EMD
}
\description{
dist1 and dist2 can be "sample-to-sample", "sample-to-prediction" or "CV-distances"
}
\details{
Calculates the Earth Mover's distance between the sample-to-prediction and the CV-distances
nearest neighbour distance distributions of a given \code{CAST::plot_geodist} output.
}
