% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMD_s2p_cv.R
\name{EMD_s2p_cv}
\alias{EMD_s2p_cv}
\title{EMD between sample-to-prediction and CV-distances}
\usage{
EMD_s2p_cv(geodist)
}
\arguments{
\item{geodist}{The output of a \code{CAST:plot_geodist()} call}
}
\value{

}
\description{
Calculates the Earth Mover's distance between the sample-to-prediction and the CV-distances
nearest neighbour distance distributions of a given \code{CAST::plot_geodist} output.
}
