% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMD_s2s_s2p.R
\name{EMD_s2s_s2p}
\alias{EMD_s2s_s2p}
\title{EMD between sample-to-sample and sample-to-prediction}
\usage{
EMD_s2s_s2p(geodist)
}
\arguments{
\item{geodist}{The output of a \code{CAST:plot_geodist()} call}
}
\value{
A numeric value.
}
\description{
Calculates the Earth Mover's distance between the sample-to-sample and the sample-to-prediction
nearest neighbour distance distributions of a given \code{CAST::plot_geodist} output.
}
