% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampled_geodist.R
\name{sampled_geodist}
\alias{sampled_geodist}
\title{Do \code{CAST::plot_geodist} with a random sample}
\usage{
sampled_geodist(
  x,
  modeldomain,
  samples,
  cvfolds = NA,
  cv_method = TRUE,
  stat = "density",
  showPlot = TRUE
)
}
\arguments{
\item{x}{sf object, training data locations}

\item{modeldomain}{sf object of the study area}

\item{samples}{size of the sample to be taken}

\item{cvfolds}{a list of CV folds}

\item{stat}{'density' or 'ecdf' for plot design}

\item{showPlot}{logical}
}
\value{
A \code{CAST::plot_geodist} list
}
\description{
Use this function like \code{CAST::plot_geodist} (only to the intended extent though), except that a random sample is taken before \code{plot_geodist} is executed. This makes the computation much faster.
}
