% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_curve.R
\name{discrete_curve}
\alias{discrete_curve}
\title{Take a \code{plot_geodist} object and put the calculated distances into equally sized bins}
\usage{
discrete_curve(gd, resolution = 100)
}
\arguments{
\item{gd}{The output of a \code{plot_geodist()} call}

\item{resolution}{numerical How many bins should the result have?}
}
\value{
A dataframe with four columns: dist, samaple-to-sample, sample-to-prediction, CV-distances
}
\description{
Take a \code{plot_geodist} object and put the calculated distances into equally sized bins
}
