% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMD.R
\name{EMD}
\alias{EMD}
\title{EMD between two distributions}
\usage{
EMD(df, dist1, dist2)
}
\arguments{
\item{df}{A dataframe, e.g. output of \code{emodis::discrete_curve()}}

\item{dist1}{The name of the first distribution (column name)}

\item{dist2}{The name of the second distributions (column name)}
}
\value{
numerical The EMD
}
\description{
dist1 and dist2 can be "sample-to-sample", "sample-to-prediction" or "CV-distances"
}
\details{
Calculates the Earth Mover's distance between two given nearest neighbour distance
distributions contained in a dataframe.
}
\examples{
\dontrun{
library(emodis)
load(system.file("extdata", "data", "mask.Rdata", package="emodis"))
# load a single example file
coordsname <- paste0(sprintf("\%03d", 1), "_coords.Rdata")
coords_file <- file.path(system.file("extdata",
"samples", package="emodis"), "clusterGapped", coordsname)
load(coords_file)
dist <- CAST::plot_geodist(pts, mask)
dist_discrete <- discrete_curve(dist)
# calculate distance between the curves
EMD(dist_discrete, "sample-to-sample", "sample-to-prediction")
}
}
